Further work on refactoring
