+++ rascontrol/__init__.py
@@ -1,1 +1,4 @@
-from rascontrol import *
+# Constants
+from rascontrol.rascontrol import SimpleXS, SimpleCulvert, SimpleBridge, SimpleMO, SimpleIS, SimpleLS
+from rascontrol.rascontrol import WSEL, MIN_CH_EL, STA_WS_LFT, STA_WS_RGT, FROUDE_CHL, FROUDE_XS, Q_WEIR, \
+    Q_CULVERT_GROUP, Q_CULVERT_TOT, WSUS, WSDS
