+++ rascontrol/exceptions/file_handling.py
@@ -1,0 +1,6 @@
+class NoOutputFile(Exception):
+    pass
+
+
+class FileNotFound(Exception):
+    pass
