+++ rascontrol/profile.py
@@ -1,0 +1,8 @@
+class Profile(object):
+    def __init__(self, name, code, rc):
+        self.name = name  # Profile name, string
+        self.code = code  # Profile code, int - these start at 1, not 0
+        self.rc = rc   # RasController object
+
+    def __repr__(self):
+        return 'Profile name = "'+self.name + '"/Profile code = "' + str(self.code)+'"'
